backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
numpy
joblib
google-genai
python-dotenv
pydantic
